Add files via upload
PCA <- When we have too many data sets we always get confused with the idea that most of them are corelated and PCA helps us in dealing with such situation.
Basically its a method of extracting important variables from a set of large no. variables in a dataset. It is even more powerful when dealing with 3 or higher dimensional data.

PCA can only be applied on numerical data so if our data set is having categorical then we should convert them into numerical data.

Some basic data cleaning is also important before implementing the technique to improve its accuracy.